requirements.txt
Flask~=3.0
flask_restx
flask-cors
python-keycloak
python_jwt